// src/models/operacion.model.ts
export class OperacionModel {
  Id_Operacion: any;
  Id_OperacionPrincipal: Number;
  TipoOrigen: String;
  CodigoOperacion: String;
  DescripcionOperacion: String;
  FechaOperacion: Date;
  ResponsableGiro: String;
  RUCResponsableGiro: String;
  Solicitud: Number;
  TipoSR: Number;
  Id_SR: Number;
  Id_DetalleSR: Number;
  DescripcionDetalle: String;
  TC: Number;
  Codigo_MedioPago: String;
  ChequeSoles: String;
  ChequeDolares: String;
  Codigo_TipoDocumento: String;
  Codigo_TipoRegistro: String;
  SerieComprobante: String;
  NumeroComprobante: String;
  FechaComprobante: Date;
  RUCAuxiliar: String;
  RazonSocial: String;
  Codigo_TipoDocumentoIdentidad: String;
  Codigo_PlanProyecto: String;
  Codigo_PlanCuenta: String;
  Nombre_PlanCuenta: String;
  MontoSoles: Number;
  MontoDolares: Number;
  DebeSoles: Number;
  HaberSoles: Number;
  DebeDolares: Number;
  HaberDolares: Number;
  C1: Number;
  C2: Number;
  C3: Number;
  C4: Number;
  C5: Number;
  C6: Number;
  C7: Number;
  C8: Number;
  C9: Number;
  C10: Number;
  V1: Number;
  V2: Number;
  V3: Number;
  V4: Number;
  V5: Number;
  V6: Number;
  V7: Number;
  Mes: String;
  Ano: Number;
  Codigo_Proyecto: String;
  Anulado: Boolean;
  Saldo: Number;
  CuentaPendiente: Number;
  TipoSaldo: Number;

  constructor() {
    this.Id_Operacion = null;
    this.Id_OperacionPrincipal = 0;
    this.TipoOrigen = '';
    this.CodigoOperacion = '';
    this.DescripcionOperacion = '';
    this.FechaOperacion = new Date();
    this.ResponsableGiro = '';
    this.RUCResponsableGiro = '';
    this.Solicitud = 0;
    this.TipoSR = 0;
    this.Id_SR = 0;
    this.Id_DetalleSR = 0;
    this.DescripcionDetalle = '';
    this.TC = 0;
    this.Codigo_MedioPago = '';
    this.ChequeSoles = '';
    this.ChequeDolares = '';
    this.Codigo_TipoDocumento = '';
    this.Codigo_TipoRegistro = '';
    this.SerieComprobante = '';
    this.NumeroComprobante = '';
    this.FechaComprobante = new Date();
    this.RUCAuxiliar = '';
    this.RazonSocial = '';
    this.Codigo_TipoDocumentoIdentidad = '';
    this.Codigo_PlanProyecto = '';
    this.Codigo_PlanCuenta = '';
    this.Nombre_PlanCuenta = '';
    this.MontoSoles = 0;
    this.MontoDolares = 0;
    this.DebeSoles = 0;
    this.HaberSoles = 0;
    this.DebeDolares = 0;
    this.HaberDolares = 0;
    this.C1 = 0;
    this.C2 = 0;
    this.C3 = 0;
    this.C4 = 0;
    this.C5 = 0;
    this.C6 = 0;
    this.C7 = 0;
    this.C8 = 0;
    this.C9 = 0;
    this.C10 = 0;
    this.V1 = 0;
    this.V2 = 0;
    this.V3 = 0;
    this.V4 = 0;
    this.V5 = 0;
    this.V6 = 0;
    this.V7 = 0;
    this.Mes = '';
    this.Ano = 0;
    this.Codigo_Proyecto = '';
    this.Anulado = false;
    this.Saldo = 0;
    this.CuentaPendiente = 0;
    this.TipoSaldo = 0;
  }
}